requirements.txt
ffmpeg_python==0.2.0
Flask==3.0.3
pywebview==5.1